support for fast-forward merges with a comment of '/fast-forward' on PRs
The default behaviour of merges in the github web interface is to make a new commit which means that the tag is not associated with the head of master,
This behaviour is unexpected and wrong and leads to tagged master commits not being released as expected.

Another alternative is to merge manually and push to 'master' github will close the PR as if it had been done through the web interface.